Make the kosmo-* layer pass the CI clippy gate (-D warnings)
The toolchain pins channel = "stable", which now floats to rustc/clippy
1.94; newer lints (added since the substrate was written against an older
stable) fire as errors under the CI RUSTFLAGS="-D warnings" gate, so the
substrate's clippy job was effectively red. Fix every violation across all
19 kosmo-* crates and 5 tools so `cargo clippy --all-targets -D warnings`
is clean.

Changes (all behavior-preserving):
- clippy.toml: raise too-many-arguments-threshold to 12 — the substrate is
  built from content-addressed records whose `new(...)` takes one arg per
  content field; 8-12 is idiomatic here. The lone 28-field aggregate
  (IntegrationRunReport::new) keeps a local #[allow] with a rationale.
- derivable_impls: manual `impl Default` -> #[derive(Default)] + #[default].
- unnecessary_map_or, manual_contains/repeat_n/pattern_char_comparison,
  needless_borrows, io_other_error, redundant_closure, owned_cow
  (Cow<Vec<T>> -> Cow<[T]>), explicit_counter_loop (-> enumerate),
  unnecessary_cast, cloned_ref_to_slice_refs (-> slice::from_ref),
  if_same_then_else (collapsed three identical PolicyGate arms).
- large_enum_variant on the TUI AppPhase: #[allow] (single live value).

Verified: full kosmo clippy gate EXIT 0; 1083 substrate tests pass, 0
failures; 147/147 eval scenarios pass.

NOTE: `cargo fmt --check` is still red across the kosmo-* layer (never
fmt'd; pse-* is clean) — handled separately.

+++ tools/kosmo-substrate/src/main.rs
@@ -211,7 +211,7 @@ fn hex8(d: &kosmo_core::Digest) -> String {
 fn q16_fmt(raw: i64) -> String {
     // Display as decimal fraction with 4 places: raw / 65536.0
     let i = raw / 65536;
-    let f = ((raw.unsigned_abs() % 65536) * 10000 / 65536) as u64;
+    let f = (raw.unsigned_abs() % 65536) * 10000 / 65536;
     format!("{}.{:04}", i, f)
 }
